Move object structured cloning code into JSC https://bugs.webkit.org/show_bug.cgi?id=319192 rdar://182042710
Reviewed by Yusuke Suzuki.

This patch moves the final remaining structured cloning code into JSC.
The object cloning code is what contains the state machine and is most
of the logic around cloning. Fortunately there's nothing in the W3C spec
that requires recursive cloning so the logic can all cleanly move into
JSC.

No new tests, no behavior change. Covered by existing tests.

Canonical link: https://commits.webkit.org/317008@main
